+++ components/ProblemStatement.js
@@ -1,9 +1,9 @@
 import tw, { styled } from 'twin.macro'
 
-function Aside({ title, children, className }) {
+function ProblemStatement({ children, className }) {
   return (
     <aside className={className}>
-      <Title>{title}</Title>
+      <Title>The Problem</Title>
       <Content>{children}</Content>
     </aside>
   )
@@ -12,7 +12,7 @@ function Aside({ title, children, className }) {
 /**
  * We have to wrap Aside in `styled` here so we can target it in a parent selector.
  */
-export default styled(Aside)``
+export default styled(ProblemStatement)``
 
 const Title = styled.h1`
   ${tw`font-mono bg-gray-800 shadow-md`};
