+++ components/debugger/TraverseVisitor.js
@@ -2,11 +2,16 @@ import React from 'react'
 import tw, { styled, theme } from 'twin.macro'
 import { BsPlayFill, BsPauseFill } from 'react-icons/bs'
 
-import { Tree, useSyntaxTree } from '../sandboxes/AstSandbox'
-import exec from '../../lib/exec'
-import usePlayer from '../../lib/usePlayer'
+import Button from '@/elements/Button'
+
+import exec from '@/lib/exec'
+import usePlayer from '@/lib/usePlayer'
+
+// Macros can't use absolute imports because they get stripped out
 import snapshot from '../../lib/snapshot.macro'
-import Button from '../elements/Button'
+
+import Tree from './shared/Tree'
+import useSyntaxTree from './shared/useSyntaxTree'
 
 const code = `let a = 10`
