+++ components/debugger/VisitorSandbox.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import { styled } from 'twin.macro'
 
-import TranspilerSandbox from './TranspilerSandbox'
-import LiveEditor from '../shared/LiveEditor'
-import Widget from '../Widget'
-import useBabelPlugin from './useBabelPlugin'
+import Widget from '@/elements/Widget'
+
+import LiveEditor from './shared/LiveEditor'
+import TranspilerSandbox from './shared/TranspilerSandbox'
+import useBabelPlugin from './shared/useBabelPlugin'
 
 /**
  * The VisitorSandbox component allows users to create their own Babel visitors and
