+++ components/debugger/shared/TranspilerSandbox.js
@@ -2,8 +2,9 @@ import React from 'react'
 import { HiArrowRight } from 'react-icons/hi'
 import tw, { styled, theme } from 'twin.macro'
 
-import LiveEditor from '../shared/LiveEditor'
-import CodeBlock from '../elements/CodeBlock'
+import CodeBlock from '@/elements/CodeBlock'
+
+import LiveEditor from './LiveEditor'
 import useBabelPlugin from './useBabelPlugin'
 
 const starterCode = `function sum(arr) {
