+++ components/debugger/shared/useSyntaxTree.js
