--- components/sandboxes/AstSandbox/index.js
@@ -1,4 +1,0 @@
-import Tree from './Tree'
-import useSyntaxTree from './useSyntaxTree'
-
-export { Tree, useSyntaxTree }
