+++ components/sliding-window/Dictionary.js
