+++ components/sliding-window/Figure.js
