+++ components/sliding-window/SlidingWindow.js
