+++ components/sliding-window/figures/FrequencyComparison.js
