+++ components/sliding-window/figures/SmallestSubarrayNaive.js
@@ -1,18 +1,17 @@
 import { motion, AnimateSharedLayout } from 'framer-motion'
 import { BsTriangleFill } from 'react-icons/bs'
-import Item from '../shared/Item'
 
-export default function Quadratic({ state }) {
+import Item from '../Item'
+
+export default function SmallestSubarrayNaive({ state }) {
   return (
-    <div className="flex flex-col items-start px-4 overflow-x-scroll md:items-center md:overflow-x-hidden">
-      <div className="flex justify-start pb-4">
+    <div className="flex flex-col px-4">
+      <div className="flex justify-center pb-4">
         <AnimateSharedLayout>
-          {state.chocolates.map((item, index) => (
+          {state.arr.map((item, index) => (
             <Item
               key={index}
-              active={
-                state.__done || (index >= state.i && index <= state.i + state.j)
-              }
+              active={state.__done || (index >= state.i && index <= state.j)}
               className="relative"
             >
               {item}
@@ -32,11 +31,10 @@ export default function Quadratic({ state }) {
         </AnimateSharedLayout>
       </div>
       <code className="block w-full mt-4 font-mono text-center whitespace-nowrap">
-        {JSON.stringify(
-          state.result.map((num) => Number(num.toFixed(2))),
-          null,
-          2
-        )}
+        s = {state.s}
+      </code>
+      <code className="block w-full font-mono text-center whitespace-nowrap">
+        min = {JSON.stringify(state.min, null, 2)}
       </code>
     </div>
   )
