+++ components/sliding-window/figures/Window.js
@@ -1,12 +1,12 @@
-import Item from '../shared/Item'
-import SlidingWindow from './shared/SlidingWindow'
+import Item from '../Item'
+import SlidingWindow from '../SlidingWindow'
 
 export default function Window({ state }) {
   const [start, end] = state.window
   return (
     <>
       <div className="flex justify-center">
-        <div className="flex justify-center items-center relative">
+        <div className="relative flex items-center justify-center">
           <SlidingWindow start={start} end={end - 1} />
           {state.arr.map((item, index) => (
             <Item key={item} active={index >= start && index < end}>
@@ -15,7 +15,7 @@ export default function Window({ state }) {
           ))}
         </div>
       </div>
-      <p className="w-full text-center mt-16 font-mono">sum: {state.sum}</p>
+      <p className="w-full mt-16 font-mono text-center">sum: {state.sum}</p>
     </>
   )
 }
