+++ components/sliding-window/figures/index.js
