+++ jsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@/components/*": ["components/*"],
+      "@/elements/*": ["elements/*"],
+      "@/lib/*": ["lib/*"]
+    }
+  }
+}
