+++ pages/debugger.mdx
@@ -4,15 +4,16 @@ blurb: 'A crash course in writing your own Babel plugin.'
 publishDate: '2021-05-07'
 ---
 
-import Aside from '../components/Aside'
-import Callout from '../components/Callout'
-import Widget from '../components/Widget'
-import Video from '../components/shared/Video'
-import EmbeddedSandbox from '../components/debugger/EmbeddedSandbox'
-import EmbeddedTranspiler from '../components/debugger/EmbeddedTranspiler'
-import VisitorSandbox from '../components/debugger/VisitorSandbox'
-import BabelPipeline from '../components/debugger/BabelPipeline'
-import TraverseVisitor from '../components/debugger/TraverseVisitor'
+import ProblemStatement from '@/components/ProblemStatement'
+import Widget from '@/elements/Widget'
+import Video from '@/elements/Video'
+
+import TreeSandbox from '@/components/debugger/TreeSandbox'
+import EmbeddedTranspiler from '@/components/debugger/EmbeddedTranspiler'
+import VisitorSandbox from '@/components/debugger/VisitorSandbox'
+import BabelPipeline from '@/components/debugger/BabelPipeline'
+import TraverseVisitor from '@/components/debugger/TraverseVisitor'
+
 import 'twin.macro'
 
 A couple of months ago, I shipped [Playground](https://playground.narendras.vercel.app/) — a web-based JavaScript debugger that lets you write JS code and attach breakpoints using the `debugger` statement. It was the first personal project that I saw through to completion, and it was an incredible opportunity for me to dip my toes in web technologies I've never used before.
@@ -62,11 +63,11 @@ This works totally fine for small and one-of functions, but it doesn't scale too
 
 This brings us to our problem:
 
-<Aside title="The Problem">
+<ProblemStatement>
   
 How do we automatically record the values of variables inside a function?
 
-</Aside>
+</ProblemStatement>
 
 ## Problem 1: Where to put it?
 
@@ -119,7 +120,7 @@ sum([1, 2, 3])
 
 ASTs always start with the entire source code as the root node of the tree:
 
-<EmbeddedSandbox
+<TreeSandbox
   initialCode={`const variables = []\n
 function sum(arr) {
   let sum = 0
@@ -137,7 +138,7 @@ sum([1, 2, 3])`}
 
 Next, we break down this source code into smaller nodes based on the code structure it represents. Here, our code consists of three parts — declaring the variable at the top, defining a function in the middle, and calling the function at the end. So let's add three more nodes as child nodes:
 
-<EmbeddedSandbox
+<TreeSandbox
   initialCode={`const variables = []\n
 function sum(arr) {
   let sum = 0
@@ -161,11 +162,11 @@ But our original problem still holds — don't we have to manipulate the code st
 
 ## The Babel APIs
 
-<Aside title="The Problem">
+<ProblemStatement>
   
 How do we take our original code string and turn it into an AST? I talked briefly about what ASTs are, but I never solved the actual problem — we still have to deal with strings anyway!
 
-</Aside>
+</ProblemStatement>
 
 To solve this problem, we're going to take advantage of [the Babel library](https://babeljs.io/). In particular, we're going to let Babel do the code parsing for us so that we can focus on manipulating the AST.
 
@@ -177,11 +178,11 @@ Fundamentally, Babel is a JavaScript compiler — it's a program that takes in J
 
 The AST that we saw in the previous section was an AST generated by Babel but with a bunch of properties cut out. In truth, Babel's ASTs are a lot more expressive. For example, here's the 'raw' AST for a function call with all the original properties shown:
 
-<EmbeddedSandbox initialCode={`helloWorld()`} hideEditor showProps />
+<TreeSandbox initialCode={`helloWorld()`} hideEditor showProps />
 
 If you want to see what kind of AST Babel generates for other pieces of code, play around with the sandbox here:
 
-<EmbeddedSandbox
+<TreeSandbox
   initialCode={`function hello() {
   console.log('hello world')
 }`}
@@ -211,11 +212,11 @@ Most, if not all, of the logic of modifying the AST lies in this visitor object 
 
 ## Be My Guest
 
-<Aside title="The Problem">
+<ProblemStatement>
   
 How do you modify an AST?
 
-</Aside>
+</ProblemStatement>
 
 To transform an AST, you have to _traverse_ through the tree, modifying each node one at a time as you visit them. A visitor is used to describe _how_ different node types are modified as you visit each one. Here's how it works.
 
@@ -281,7 +282,7 @@ I've embedded a visitor and AST playground for you to use. Make changes to the c
 })`}
 />
 
-<EmbeddedSandbox
+<TreeSandbox
   initialCode={`function sum(arr) {
   let sum = 0
   for (const num of arr) {
@@ -299,7 +300,7 @@ I've embedded a visitor and AST playground for you to use. Make changes to the c
 
 Let's work through this together. The first thing we have to do is figure out which node type(s) we should target in the visitor. Thankfully, there's a dedicated node type specifically for the `debugger` statement we're interested in:
 
-<EmbeddedSandbox
+<TreeSandbox
   initialCode={`const a = b
 debugger`}
   depth={2}
@@ -337,7 +338,7 @@ A **builder** is a helper function used to create AST nodes. When we use builder
 
 Here's what the mini AST could look like:
 
-<EmbeddedSandbox
+<TreeSandbox
   initialCode={`_variables.push({ a })`}
   depth={2}
   hideEditor
