+++ pages/sliding-window.mdx
@@ -4,13 +4,15 @@ blurb: 'Chocolates and a way to make your array algorithms a bit faster.'
 publishDate: '2021-03-02'
 ---
 
-import Visualizer from '../components/Visualizer'
-import Callout from '../components/Callout'
-import Figure from '../components/Figure'
-import Item from '../components/shared/Item'
-import SlidingWindow from '../components/sliding-window/shared/SlidingWindow'
-import Dictionary from '../components/sliding-window/shared/Dictionary'
-import * as Figures from '../components/sliding-window'
+import Visualizer from '@/components/sliding-window/Visualizer'
+import Figure from '@/components/sliding-window/Figure'
+import Item from '@/components/sliding-window/Item'
+import SlidingWindow from '@/components/sliding-window/SlidingWindow'
+import Dictionary from '@/components/sliding-window/Dictionary'
+import * as Figures from '@/components/sliding-window/figures'
+
+import Callout from '@/elements/Callout'
+
 import snapshot from '../lib/snapshot.macro'
 import { motion } from 'framer-motion'
 import 'twin.macro'
